backend: logger: Move TestLogErrorTypes to internal test
